GFAL2Base class with inheritance classes
