+++ Resources/Storage/GFAL2StorageBase.py
@@ -42,7 +42,7 @@ class GFAL2StorageBase( StorageBase ):
 #     os.environ['GLOBUS_THREAD_MODEL'] = "pthread"
     StorageBase.__init__( self, storageName, parameters )
 
-    self.log = gLogger.getSubLogger( "SRM2V2Storage", True )
+    self.log = gLogger.getSubLogger( "GFAL2StorageBase", True )
 
     # Different levels or verbosity:
     # gfal2.verbose_level.normal,
@@ -54,7 +54,7 @@ class GFAL2StorageBase( StorageBase ):
     if dlevel == 'DEBUG':
       gfal2.set_verbose( gfal2.verbose_level.trace )
 
-
+    gfal2.set_verbose( gfal2.verbose_level.trace )
     self.isok = True
 
     # # gfal2 API
@@ -65,7 +65,52 @@ class GFAL2StorageBase( StorageBase ):
     self.protocol = parameters['Protocol']
     self.spaceToken = parameters['SpaceToken']
 
-    # # init base class
+
+    # #stage limit - 12h
+    self.stageTimeout = gConfig.getValue( '/Resources/StorageElements/StageTimeout', 12 * 60 * 60 )  # gConfig -> [get] ConfigurationClient()
+    # # 1 file timeout
+    self.fileTimeout = gConfig.getValue( '/Resources/StorageElements/FileTimeout', 30 )
+    # # nb of surls per gfal2 call
+    self.filesPerCall = gConfig.getValue( '/Resources/StorageElements/FilesPerCall', 20 )
+    # # gfal2 timeout
+    self.gfal2Timeout = gConfig.getValue( "/Resources/StorageElements/GFAL_Timeout", 100 )
+    # # gfal2 long timeout
+    self.gfal2LongTimeOut = gConfig.getValue( "/Resources/StorageElements/GFAL_LongTimeout", 1200 )
+    # # gfal2 retry on errno.ECONN
+    self.gfal2Retry = gConfig.getValue( "/Resources/StorageElements/GFAL_Retry", 3 )
+
+
+    # # set checksum type, by default this is 0 (GFAL_CKSM_NONE)
+    self.checksumType = gConfig.getValue( "/Resources/StorageElements/ChecksumType", 0 )
+    # enum gfal_cksm_type, all in lcg_util
+    #   GFAL_CKSM_NONE = 0,
+    #   GFAL_CKSM_CRC32,
+    #   GFAL_CKSM_ADLER32,
+    #   GFAL_CKSM_MD5,
+    #   GFAL_CKSM_SHA1
+    # GFAL_CKSM_NULL = 0
+    self.checksumTypes = { None : 0, "CRC32" : 1, "ADLER32" : 2,
+                           "MD5" : 3, "SHA1" : 4, "NONE" : 0, "NULL" : 0 }
+    if self.checksumType:
+      if str( self.checksumType ).upper() in self.checksumTypes:
+        gLogger.debug( "GFAL2StorageBase: will use %s checksum check" % self.checksumType )
+        self.checksumType = self.checksumTypes[ self.checksumType.upper() ]
+      else:
+        gLogger.warn( "GFAL2StorageBase: unknown checksum type %s, checksum check disabled" )
+        # # GFAL_CKSM_NONE
+        self.checksumType = 0
+    else:
+      # # invert and get name
+      self.log.debug( "GFAL2StorageBase: will use %s checksum" % dict( zip( self.checksumTypes.values(),
+                                                                     self.checksumTypes.keys() ) )[self.checksumType] )
+    self.voName = None
+    ret = getProxyInfo( disableVOMS = True )
+    if ret['OK'] and 'group' in ret['Value']:
+      self.voName = getVOForGroup( ret['Value']['group'] )
+    self.defaultLocalProtocols = gConfig.getValue( '/Resources/StorageElements/DefaultProtocols', [] )
+
+    self.MAX_SINGLE_STREAM_SIZE = 1024 * 1024 * 10  # 10 MB ???
+    self.MIN_BANDWIDTH = 0.5 * ( 1024 * 1024 )  # 0.5 MB/s ???
 
 
 
@@ -1804,115 +1849,6 @@ class GFAL2StorageBase( StorageBase ):
 #     """
 #     return S_OK()
 ##################################################################
-
-
-
-  def getTransportURL( self, path, protocols = False ):
-    """ obtain the tURLs for the supplied path and protocols
-
-    :param self: self reference
-    :param str path: path on storage
-    :param mixed protocols: protocols to use
-    :returns Failed dict {path : error message}
-             Successful dict {path : transport url}
-             S_ERROR in case of argument problems
-    """
-    res = checkArgumentFormat( path )
-    if not res['OK']:
-      return res
-    urls = res['Value']
-
-    self.log.debug( 'GFAL2StorageBase.getTransportURL: Attempting to retrieve tURL for %s paths' % len( urls ) )
-
-    failed = {}
-    successful = {}
-
-    if not protocols:
-      protocols = self.__getProtocols()
-      if not protocols['OK']:
-        return protocols
-      listProtocols = protocols['Value']
-    elif type( protocols ) == StringType:
-      listProtocols = [protocols]
-    elif type( protocols ) == ListType:
-      listProtocols = protocols
-    else:
-      return S_ERROR( "getTransportURL: Must supply desired protocols to this plug-in." )
-
-    for url in urls:
-      res = self.__getSingleTransportURL( url, listProtocols )
-      self.log.debug( 'res = %s' % res )
-
-      if not res['OK']:
-        failed[url] = res['Message']
-      else:
-        successful[url] = res['Value']
-
-    return S_OK( { 'Failed' : failed, 'Successful' : successful } )
-
-
-
-  def __getSingleTransportURL( self, path, protocols = False ):
-    """ Get the tURL from path with getxattr from gfal2
-
-    :param self: self reference
-    :param str path: path on the storage
-    :returns S_OK( Transport_URL ) in case of success
-             S_ERROR( errStr ) in case of a failure
-    """
-    self.log.debug( 'GFAL2StorageBase.__getSingleTransportURL: trying to retrieve tURL for %s' % path )
-    if protocols:
-      res = self.__getExtendedAttributes( path, protocols )
-    else:
-      res = self.__getExtendedAttributes( path )
-    if res['OK']:
-      attributeDict = res['Value']
-      # 'user.replicas' is the extended attribute we are interested in
-      if 'user.replicas' in attributeDict.keys():
-        turl = attributeDict['user.replicas']
-        return S_OK( turl )
-      else:
-        errStr = 'GFAL2StorageBase.__getSingleTransportURL: Extended attribute tURL is not set.'
-        self.log.debug( errStr )
-        return S_ERROR( errStr )
-    else:
-      errStr = 'GFAL2StorageBase.__getSingleTransportURL: %s' % res['Message']
-      return S_ERROR( errStr )
-
-
-
-  def __getProtocols( self ):
-    """ returns list of protocols to use at a given site
-
-    :warn: priority is given to a protocols list defined in the CS
-
-    :param self: self reference
-    """
-    sections = gConfig.getSections( '/Resources/StorageElements/%s/' % ( self.name ) )
-    self.log.debug( "GFAL2StorageBase.__getProtocols: Trying to get protocols for storage %s." % self.name )
-    if not sections['OK']:
-      return sections
-
-    protocolsList = []
-    for section in sections['Value']:
-      path = '/Resources/StorageElements/%s/%s/ProtocolName' % ( self.name, section )
-      if gConfig.getValue( path, '' ) == self.protocol:
-        protPath = '/Resources/StorageElements/%s/%s/ProtocolsList' % ( self.name, section )
-        siteProtocols = gConfig.getValue( protPath, [] )
-        if siteProtocols:
-          self.log.debug( 'GFAL2StorageBase.__getProtocols: Found SE protocols list to override defaults:', ', '.join( siteProtocols, ) )
-          protocolsList = siteProtocols
-
-    if not protocolsList:
-      self.log.debug( "GFAL2StorageBase.__getProtocols: No protocols provided, using the default protocols." )
-      protocolsList = gConfig.getValue( '/Resources/StorageElements/DefaultProtocols', [] )
-
-    # if there is even no default protocol
-    if not protocolsList:
-      return S_ERROR( "GFAL2StorageBase.__getProtocols: No local protocols defined and no defaults found." )
-
-    return S_OK( protocolsList )
-
 
 
   def __getExtendedAttributes( self, path, protocols = False ):
